+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   * Service added to optimize handling of EXTCONF values.
   * Configuration option for cookie name added.
   * Configuration option for cookie lifetime added.
+  * Configuration option for brute force sleep duration added.
 * Changed
   * Password field changed to password type.
   * Configuration variable for update URI moved to _ext_localconf.php_.